+++ src/main/scala/org/broadinstitute/dsde/workbench/sam/db/tables/FlatGroupMemberTable.scala
@@ -1,9 +1,8 @@
 package org.broadinstitute.dsde.workbench.sam.db.tables
 
-import java.sql.ResultSet
-
 import org.broadinstitute.dsde.workbench.model.WorkbenchUserId
-import org.broadinstitute.dsde.workbench.sam.db.{DatabaseArray, DatabaseKey, SamTypeBinders}
+import org.broadinstitute.dsde.workbench.sam.db.SamTypeBinders._
+import org.broadinstitute.dsde.workbench.sam.db.{DatabaseArray, DatabaseKey}
 import scalikejdbc._
 
 final case class FlatGroupMemberPK(value: Long) extends DatabaseKey
@@ -24,28 +23,9 @@ final case class FlatGroupMemberRecord(id: FlatGroupMemberPK,
                                        memberGroupId: Option[GroupPK],
                                        groupMembershipPath: FlatGroupMembershipPath,
                                        lastGroupMembershipElement: Option[GroupPK])
-// TODO move to SamTypeBinders?
-object TempTypeBinders {
-  implicit val flatGroupMemberPKTypeBinder: TypeBinder[FlatGroupMemberPK] = new TypeBinder[FlatGroupMemberPK] {
-    def apply(rs: ResultSet, label: String): FlatGroupMemberPK = FlatGroupMemberPK(rs.getLong(label))
-    def apply(rs: ResultSet, index: Int): FlatGroupMemberPK = FlatGroupMemberPK(rs.getLong(index))
-  }
-
-  implicit val flatGroupMembershipPathPKTypeBinder: TypeBinder[FlatGroupMembershipPath] = new TypeBinder[FlatGroupMembershipPath] {
-    def apply(rs: ResultSet, label: String): FlatGroupMembershipPath = {
-      FlatGroupMembershipPath(rs.getArray(label).getArray.asInstanceOf[Array[java.lang.Long]].map(_.longValue()).toList.map(GroupPK))
-    }
-    def apply(rs: ResultSet, index: Int): FlatGroupMembershipPath = {
-      FlatGroupMembershipPath(rs.getArray(index).getArray.asInstanceOf[Array[java.lang.Long]].map(_.longValue()).toList.map(GroupPK))
-    }
-  }
-}
-
 object FlatGroupMemberTable extends SQLSyntaxSupportWithDefaultSamDB[FlatGroupMemberRecord] {
   override def tableName: String = "SAM_GROUP_MEMBER_FLAT"
 
-  import TempTypeBinders._
-  import SamTypeBinders._
   def apply(e: ResultName[FlatGroupMemberRecord])(rs: WrappedResultSet): FlatGroupMemberRecord = FlatGroupMemberRecord(
     rs.get(e.id),
     rs.get(e.groupId),
